LayoutTests/fast/sandbox/mac/sandbox-mach-lookup-expected.txt

Regression tests for mach lookup sandbox changes on macOS

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.hasSandboxMachLookupAccessToGlobalName("com.apple.WebKit.WebContent", "com.apple.cfprefsd.agent") is false
PASS internals.hasSandboxMachLookupAccessToGlobalName("com.apple.WebKit.WebContent", "com.apple.system.logger") is false
